working on notifications

+++ app/controllers/signups_controller.rb
@@ -14,21 +14,6 @@ class SignupsController < ApplicationController
         @signup.user = current_user
         @signup.campaign_id = params[:campaign_id]
         @signup.save
-        if @signup.save
-            players = Signup.where(campaign_id: @campaign.id).length
-            p '*' *100
-            p players
-            p @campaign.size
-            p @campaign.full
-            p '*' *100
-            if (players == @campaign.size)
-                @campaign.full = true
-                @campaign.save
-            end
-            p '*' *100
-            p @campaign.full
-            p '*' *100
-        end
         redirect_to campaign_path(params[:campaign_id])
     end
 
@@ -36,20 +21,35 @@ class SignupsController < ApplicationController
         @signup = Signup.find(params[:id])
         @signup.approved = true
         @signup.save
+        if @signup.save
+            players = Signup.where(campaign_id: @campaign.id).length
+            if (players == @campaign.size)
+                @campaign.full = true
+                @campaign.save
+            end
+        end
         redirect_to notifications_path
     end
 
     def destroy
-        @campaign = Campaign.find(params[:id])
-        @signup = Signup.where(campaign_id: @campaign.id).where(user_id: current_user)
-        p '*' * 100
-        p @signup
-        @signup.destroy_all
-        if @signup.destroy_all
-            players = Signup.where(campaign_id: @campaign.id).length
-            @campaign.full = false
+        if Campaign.find(params[:campaign_id])
+            @signup = Signup.find(params[:id])
+            @campaign = Campaign.where(id: @signup.campaign_id)
+            if @signup.destroy
+                players = Signup.where(campaign_id: @campaign.id).length
+                @campaign.full = false
+            end
+            redirect_to notifications_path
+        else
+            @campaign = Campaign.find(params[:id])
+            @signup = Signup.where(campaign_id: @campaign.id).where(user_id: current_user)
+            @signup.destroy_all
+            if @signup.destroy_all
+                players = Signup.where(campaign_id: @campaign.id).length
+                @campaign.full = false
+            end
+            redirect_to campaigns_path
         end
-        redirect_to campaigns_path
     end
 
 end
